msservice/requirements.txt
Django
django-cors-headers
django-rest-swagger==0.3.8
djangorestframework
djangorestframework-jwt
python-dateutil
PyJWT
PyYAML
wheel
celery
redis
psycopg2
requests
rethinkdb
